requirements.txt
pandas
numpy
matplotlib
biopython
tk
requests